main.py
"""
AIOps Agent — Main Orchestrator
================================
Entry point for the autonomous operations agent.
Coordinates all layers: ingestion → detection → drift monitoring → decision → action.

Phase progress:
  v0.1 ✅ Skeleton
  v0.2 🔜 Telemetry simulator
  v0.3 🔜 Anomaly detector
  v0.4 🔜 Evidently drift monitor
  v0.5 🔜 Policy engine
  v0.6 🔜 Feedback loop
"""

from dotenv import load_dotenv
load_dotenv()


def run_agent():
    print("=" * 60)
    print("  🤖 AIOps Agent — Starting Up")
    print("=" * 60)

    # Phase 2: Ingest / simulate telemetry
    # from data_ingestion.simulator import TelemetrySimulator
    # sim = TelemetrySimulator()
    # reference_df, current_df = sim.generate()

    # Phase 3: Anomaly detection
    # from anomaly_detection.detector import AnomalyDetector
    # detector = AnomalyDetector()
    # detector.fit(reference_df)
    # scores = detector.predict(current_df)

    # Phase 4: Drift monitoring
    # from drift_monitor.evidently_runner import DriftMonitor
    # monitor = DriftMonitor()
    # drift_result = monitor.run(reference_df, current_df)

    # Phase 5: Policy + actions
    # from policy_engine.policy import PolicyEngine
    # engine = PolicyEngine()
    # decision = engine.decide(scores, drift_result)

    print("\n  ⚙️  Agent skeleton initialised.")
    print("  📦  Phase 1 (v0.1-skeleton) complete.")
    print("\n  Next: implement data_ingestion/simulator.py (Phase 2)")
    print("=" * 60)


if __name__ == "__main__":
    run_agent()
