requirements.txt
# AIOps Agent — Python Dependencies
# Install with: pip install -r requirements.txt

# Core data
pandas>=2.0.0
numpy>=1.24.0

# Anomaly detection
scikit-learn>=1.3.0
pyod>=1.1.0

# Drift monitoring
evidently>=0.4.0

# Visualization
matplotlib>=3.7.0

# Environment & config
python-dotenv>=1.0.0

# HTTP (for future API / webhook actions)
requests>=2.31.0

# Optional: LLM reasoning layer
# anthropic>=0.25.0     # Uncomment to enable Claude API
# openai>=1.0.0         # Uncomment to enable OpenAI API